SPEC: updated version of rtl-sdr pulled
Until now, we were using a manually packaged/tarballed version of rtl-sdr
because the rtl-sdr maintainer unfortunately does not know how
to properly package and release his rtl tarballs (tarball name
not the same as uncompressed root folder, etc...). Somehow in the
process of using an actual .tar.gz in the /deps folder, we realized
that rtl-sdr version that we were using was not latest 0.5.3 as
expected, but version 0.5.0.

This change fixes that and uses now librtlsdr-0.5.3 in the /deps